remove redundant functions from  presto client
